Undo/Redo: movable line; plot function; connecting line; show squares
These all follow the same "toggle" pattern.

Also update previously-changed togglePlottedCount to follow the same
pattern, instead of the mistakenly-introduced "triggerToggleX" pattern
we had before.

Note: the one toggle that looks like it ought to belong here, but is actually
more complicated, is toggleInterceptLocked. When we turn on interceptLocked,
it changes the line. Undo should put the line back to where it was before,
instead of merely toggling the lock state. The next commit will address this.

[#96680608]