Support the ability to insert a new case at a specified index
Previously, when a case was created, the new case would be added to
the end of the collection's caseRecords array, and would be given an
index greater than the previous case.

Now we can pass an `index` property in the properties object. If we
do this, the new case will be inserted in the caseRecords array in the
appropriate place, and given an index that reflects this position.

[#96680770]