Undo/Redo: Save game state when closing DataInteractive component
When we close a DI component, we asynchronously try to save the game state
before we actually close the component. When undoing, we then restore that
game state after the component is re-opened.

We use the new saveGameState method from 3e64428e to save the state.

We also need to be able to prevent dirtying the document when the component
finally closes, as this action will actually occur after the command has
finished executing, and so would otherwise clear the undo history.

[#96701234]