Don't delete child cases twice
When we select a parent (which selects the children) and hit "delete",
we call `deleteCaseAndChildren` on all selected cases.

When the parent is deleted, it first deletes all the children.

We then continue to try and delete the already-deleted children.

This was causing the undo-data to mistakenly think that more cases had
been deleted than actually were.

[#96680770]